Setup StartMatch packet receiving in Client

+++ Source/Demos/ClientServerGame/Client/src/public/states/CPlayingState.h
@@ -12,14 +12,14 @@ class CTestLevel;
 class CPlayingState : public GPlayingState
 {
 public:
-	Delegate<const uint32_t()> m_unidRequestNetworkId;
+	Delegate<const uint8_t()> m_unidRequestNetworkId;
 
 protected:
 private:
 	Renderer* const m_cpRenderer;
 	EventHandler* const m_cpEventHandler;
 
-	uint32_t m_uNetworkId = (uint32_t)-1;
+	uint8_t m_uNetworkId = (uint8_t)-1;
 
 public:
 	CPlayingState(Renderer* const i_cpRenderer, EventHandler* const i_cpEventHandler);
